+++ app/services/projects_for_export_service.rb
@@ -1,9 +1,4 @@
 class ProjectsForExportService
-  def conversion_projects_by_advisory_board_date(month:, year:)
-    projects = Conversion::Project.filtered_by_advisory_board_date(month, year)
-    AcademiesApiPreFetcherService.new.call!(projects)
-  end
-
   def transfer_projects_by_advisory_board_date(month:, year:)
     projects = Transfer::Project.filtered_by_advisory_board_date(month, year)
     AcademiesApiPreFetcherService.new.call!(projects)